DEV-1689 Fixes Part 3
Added suggested changes from Julien to fix where images appear and step numbering so that it's more consistent from page to page even when the .md pattern seems consistent in the code (change made in image.html).  Also a separate fix for the broken link to I had to the gateway verification file.